requirements.txt
python-dotenv
requests
azure-identity
azure-ai-projects==1.1.0b4

